Moved demography T model test fixture into shared conftest.py
